Fixed test that got redirected wrongly.
This went wrong on the Zope4 PLIP.
It looks like self.browser.post changed to not only post,
but also handle any redirects that are done afterwards.
This meant that portal_quickinstaller/installProducts got loaded as GET, which is not allowed.
So we do a GET ourselves first, so a referer is set that is then redirected to.